# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Forafix API Documentation",
        "description": "Marketplace API with Wallet, Escrow and Booking management",
        "contact": {
            "email": "support@forafix.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sessions": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "List all active user sessions and tokens",
                "operationId": "8db417cfbdb154beb6fcc56d6649967f",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sessions/{id}": {
            "delete": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Revoke a specific session or token",
                "operationId": "af1f4ecb5a27ba7214dfd7bd177b8e8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session revoked"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sessions/revoke-others": {
            "post": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Revoke all other active sessions and tokens",
                "operationId": "674b38f08b84d9553a7af39970263a11",
                "responses": {
                    "200": {
                        "description": "Sessions revoked"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/two-factor/enable": {
            "post": {
                "tags": [
                    "Security"
                ],
                "summary": "Enable 2FA (Generates Secret & QR Code)",
                "operationId": "28f075298a32079c6c2fe8b82db539d7",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/two-factor/confirm": {
            "post": {
                "tags": [
                    "Security"
                ],
                "summary": "Confirm 2FA setup with a code",
                "description": "Confirm 2FA setup.",
                "operationId": "1d69bf9c0bcaf76f1e6693ff452d76fa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Invalid Code"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/two-factor/disable": {
            "post": {
                "tags": [
                    "Security"
                ],
                "summary": "Disable 2FA",
                "operationId": "a787111df98cc5c76450c248c068459e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/two-factor/challenge": {
            "post": {
                "tags": [
                    "Security"
                ],
                "summary": "Verify 2FA code during login",
                "operationId": "46c55f91a5aead8c9a3b7fa187ca8baf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "code"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "code": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Invalid Code"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}